<test> 단일 역 조회 인수 테스트 코드 추가
