+++ src/main/java/nextstep/subway/applicaion/dto/StationResponse.java
@@ -22,6 +22,9 @@ public class StationResponse {
                 .collect(Collectors.toList());
     }
 
+    private StationResponse() {
+    }
+
     public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
